refactor: Change the package to entity Product and too change the name
