+++ src/main/java/io/shoppingcard/infrastructure/database/StorageMemory.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.micronaut.context.annotation.Factory;
-import io.shoppingcard.business.dto.Product;
+import io.shoppingcard.infrastructure.database.entity.ProductEntity;
 import jakarta.inject.Singleton;
 
 import java.io.BufferedReader;
@@ -20,7 +20,7 @@ import java.util.stream.Collectors;
 public class StorageMemory {
 
     @Singleton
-    List<Product> products() throws IOException {
+    List<ProductEntity> products() throws IOException {
         //TODO: The path is in env path
         try (InputStream is = TypeReference.class.getResourceAsStream("/json/products.json")) {
             String jsonText = new BufferedReader(
@@ -28,7 +28,7 @@ public class StorageMemory {
                     .lines()
                     .collect(Collectors.joining("\n"));
 
-            return Arrays.asList(new ObjectMapper().readValue(jsonText, Product[].class));
+            return Arrays.asList(new ObjectMapper().readValue(jsonText, ProductEntity[].class));
         } catch (JsonProcessingException e) {
             //TODO: Add an logger with error
             e.printStackTrace();
